vm generate_scripts
