+++ src/c_aci_testing/args/parameters/cplat.py
@@ -10,25 +10,25 @@ import os
 
 def parse_cplat_args(parser):
     parser.add_argument(
-    "--cplat-feed",
-    type=str,
-    default=os.getenv("CPLAT_FEED", ""),
-  )
+        "--cplat-feed",
+        type=str,
+        default=os.getenv("CPLAT_FEED", ""),
+    )
     parser.add_argument(
-    "--cplat-name",
-    type=str,
-    default=os.getenv("CPLAT_NAME", ""),
-  )
+        "--cplat-name",
+        type=str,
+        default=os.getenv("CPLAT_NAME", ""),
+    )
     parser.add_argument(
-    "--cplat-version",
-    type=str,
-    default=os.getenv("CPLAT_VERSION", ""),
-  )
+        "--cplat-version",
+        type=str,
+        default=os.getenv("CPLAT_VERSION", ""),
+    )
     parser.add_argument(
-    "--cplat-blob-name",
-    type=str,
-    default=os.getenv("CPLAT_BLOB_NAME", ""),
-  )
+        "--cplat-blob-name",
+        type=str,
+        default=os.getenv("CPLAT_BLOB_NAME", ""),
+    )
     parser.add_argument(
         "--cplat-path",
         type=str,
