+++ src/c_aci_testing/args/subparsers/vm.py
@@ -39,12 +39,22 @@ def subparse_vm(vm: argparse.ArgumentParser):
     parse_vm_size(create)
     parse_vm_zones(create)
 
+    generate_scripts = vm_subparser.add_parser("generate_scripts")
+    parse_target_path(generate_scripts)
+    generate_scripts.add_argument("output_dir", type=str)
+    parse_subscription(generate_scripts)
+    parse_resource_group(generate_scripts)
+    parse_vm_win_flavor(generate_scripts)
+    parse_registry(generate_scripts)
+    parse_repository(generate_scripts)
+    parse_tag(generate_scripts)
+    parse_runc_prefix(generate_scripts)
+
     runc = vm_subparser.add_parser("runc")
     parse_target_path(runc)
     parse_deployment_name(runc)
     parse_subscription(runc)
     parse_resource_group(runc)
-    parse_managed_identity(runc)
     parse_storage_account(runc)
     parse_vm_win_flavor(runc)
     parse_registry(runc)
@@ -96,7 +106,6 @@ def subparse_vm(vm: argparse.ArgumentParser):
     parse_deployment_name(cp_into)
     parse_subscription(cp_into)
     parse_resource_group(cp_into)
-    parse_managed_identity(cp_into)
     parse_storage_account(cp_into)
 
     exec = vm_subparser.add_parser("exec")
@@ -110,7 +119,6 @@ def subparse_vm(vm: argparse.ArgumentParser):
     parse_deployment_name(cat)
     parse_subscription(cat)
     parse_resource_group(cat)
-    parse_managed_identity(cat)
     parse_storage_account(cat)
 
     cache_cplat = vm_subparser.add_parser("cache_cplat")
